refactor: use transpile normalisation for nuxt3 too
